requirements.txt
# Core web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ML and embeddings - ONNX Runtime (lightweight!)
onnxruntime==1.23.2
onnxscript==0.5.6
transformers==4.38.2
tokenizers==0.15.2
numpy==1.26.4

# Utility libraries
pydantic==2.6.3
python-dotenv==1.0.1
requests==2.31.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0

# Development
black==24.2.0
flake8==7.0.0

# FAISS
faiss-cpu